aidevs3/w2/s02e01/requirements.txt
openai
openai-whisper
PyYAML
